feat(text-service): 完成文章模块的一些api
1.完成文章新增
2.完成文章搜索

+++ text-service/src/main/java/com/llpy/textservice/entity/ArticleGroup.java
@@ -3,7 +3,9 @@ package com.llpy.textservice.entity;
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.extension.activerecord.Model;
 import com.baomidou.mybatisplus.annotation.TableId;
+
 import java.io.Serializable;
+
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
@@ -19,14 +21,14 @@ import lombok.EqualsAndHashCode;
  */
 @Data
 @EqualsAndHashCode(callSuper = false)
-@ApiModel(value="ArticleGroup对象", description="文章分组表")
+@ApiModel(value = "ArticleGroup对象", description = "文章分组表")
 public class ArticleGroup extends Model<ArticleGroup> {
 
     private static final long serialVersionUID = 1L;
 
     @ApiModelProperty(value = "分组ID，自增主键")
-      @TableId(value = "id", type = IdType.AUTO)
-    private Long id;
+    @TableId(value = "id", type = IdType.AUTO)
+    private Integer id;
 
     @ApiModelProperty(value = "文章类型")
     private String articleType;
